+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.2.0](https://github-vivi/vivitian930/cv-builder/compare/v0.1.0...v0.2.0) (2023-07-07)
+
+
+### Features
+
+* using pydantic class ([8adca6e](https://github-vivi/vivitian930/cv-builder/commit/8adca6ee00e97658723c79d6870b08b5ef908d1a))
+
+
+### Bug Fixes
+
+* style personal info ([ee8da13](https://github-vivi/vivitian930/cv-builder/commit/ee8da137345b13c91798ef1f04d5003e3e02892e))
+
 ## 0.1.0 (2023-07-07)